[Silver III] Title: 주유소, Time: 440 ms, Memory: 40072 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -18,11 +18,11 @@ class Main {
         int[] price = new int[N];
         st = new StringTokenizer(br.readLine());
         price[0] = Integer.parseInt(st.nextToken());
-        ArrayList<int[]> arr = new ArrayList<>();
+        ArrayList<long[]> arr = new ArrayList<>();
         for(int i=1; i<N; i++){
             price[i] = Integer.parseInt(st.nextToken());
             if(price[i] <= price[i-1] || i == N-1){
-                arr.add(new int[]{price[i-1], dist[i-1]});
+                arr.add(new long[]{price[i-1], dist[i-1]});
             }else{
                 price[i] = price[i-1];
                 dist[i] = dist[i-1] + dist[i];
@@ -30,11 +30,8 @@ class Main {
         }
 
         // 거리의 누적합을 가격이 저렴한 구간이 나오기 전까지 구한다.
-        // 
-        // 1. 현재 가격보다 싼 위치를 정한다.
-        // 2. 해당 위치까지 가는 데 드는 비용만큼 주유한다.
-        int answer = 0;
-        for(int[] x : arr){
+        long answer = 0;
+        for(long[] x : arr){
             answer += x[0] * x[1];
         }
         System.out.println(answer);
